ci: set up multi-platform azure pipelines
This CI pipeline checks that the bindings work natively by running the go tests against them.
The cross-compilation checks are limited to x86 for now as they can execute on azure's amd64 machines. Other targets such as arm will be later checked used `qemu-user-static` on linux.

A set of go build tags allow to either:
- use the Go bindings when supported.
- disable them even if supported.
- stubbed by a failsafe implementation.

Only linux,amd64 and darwin,amd64 are supported. Any other target is properly stubbed, allowing the agent to compile and properly work.

+++ waf/internal/bindings/waf.go
@@ -2,6 +2,11 @@
 // Please refer to our terms for more information:
 // https://www.sqreen.io/terms.html
 
+// +build !sqreen_nowaf
+// +build !windows
+// +build amd64
+// +build linux darwin
+
 package bindings
 
 import (
@@ -14,8 +19,9 @@ import (
 	"github.com/sqreen/go-libsqreen/waf/types"
 )
 
-// #cgo CFLAGS: -I${SRCDIR}/../../../lib/include
-// #cgo linux amd64 LDFLAGS: -L${SRCDIR}/../../../lib/lib64/
+// #cgo CFLAGS: -I${SRCDIR}/../../../lib/waf/include
+// #cgo amd64,linux LDFLAGS: -L${SRCDIR}/../../../lib/waf/amd64/linux
+// #cgo amd64,darwin LDFLAGS: -L${SRCDIR}/../../../lib/waf/amd64/darwin
 // #cgo LDFLAGS: -lwaf -lstdc++
 // #include <stdlib.h>
 // #include <string.h>
@@ -100,7 +106,7 @@ func (r Rule) Run(data types.RunInput, timeout time.Duration) (action types.Acti
 	}
 	defer C.powerwaf_freeInput(dataIn, C.bool(false))
 
-	ret := C.powerwaf_runPowerWAF(r.id, dataIn, C.ulong(timeout/time.Microsecond))
+	ret := C.powerwaf_runPowerWAF(r.id, dataIn, C.size_t(timeout/time.Microsecond))
 	defer C.powerwaf_freeReturn(ret)
 
 	switch a := ret.action; a {
@@ -154,7 +160,7 @@ func valueToWAFInput(v reflect.Value) (in *C.PWArgs, err error) {
 			k := k.String()
 			key := C.CString(k)
 			defer C.free(unsafe.Pointer(key))
-			if !C.powerwaf_addToPWArgsMap(in, key, C.ulong(len(k)), *value) {
+			if !C.powerwaf_addToPWArgsMap(in, key, C.size_t(len(k)), *value) {
 				C.powerwaf_freeInput(value, C.bool(false))
 				C.powerwaf_freeInput(in, C.bool(false))
 				return nil, errors.New("could not insert a key element into a map")
@@ -187,7 +193,7 @@ func valueToWAFInput(v reflect.Value) (in *C.PWArgs, err error) {
 	case reflect.Int32:
 		fallthrough
 	case reflect.Int64:
-		arg := C.powerwaf_createInt((C.long)(v.Int()))
+		arg := C.powerwaf_createInt((C.int64_t)(v.Int()))
 		return &arg, nil
 
 	case reflect.Uint:
@@ -199,7 +205,7 @@ func valueToWAFInput(v reflect.Value) (in *C.PWArgs, err error) {
 	case reflect.Uint32:
 		fallthrough
 	case reflect.Uint64:
-		arg := C.powerwaf_createUint((C.ulong)(v.Uint()))
+		arg := C.powerwaf_createUint((C.uint64_t)(v.Uint()))
 		return &arg, nil
 	}
 }
